Add delete_agenda_by_conference_id function and update delete_conference and delete_session functions

+++ app/crud/conferences_crud.py
@@ -3,6 +3,7 @@ from sqlalchemy.orm import Session
 from datetime import datetime, date
 from .. import models
 from ..schemas import conference_schemas as schemas
+from . import agenda_crud
 from pytz import timezone
 import uuid
 
@@ -35,11 +36,16 @@ def get_conference_by_conference_uuid(db: Session, uuid: str):
     return db.query(models.Conference).filter(models.Conference.uuid == uuid).first()
 
 # delete conference by conference id
-def delete_conference(db: Session,owner_id: int, uuid: str):
+def delete_conference(db: Session, owner_id: int, uuid: str):
     conference = db.query(models.Conference).filter(models.Conference.uuid == uuid, models.Conference.owner_id == owner_id).first()
-    db.query(models.Session).filter(models.Session.conference_id == conference.id, models.Session.owner_id==owner_id).delete()
-    db.query(models.Settings).filter(models.Settings.conference_id == conference.id, models.Settings.owner_id==owner_id).delete()
-    db.query(models.Conference).filter(models.Conference.uuid == uuid, models.Conference.owner_id == owner_id).delete()
+    if conference is None:
+        return False
+    sessions = db.query(models.Session).filter(models.Session.conference_id == conference.id, models.Session.owner_id == owner_id)
+    for session in sessions:
+        db.delete(session)
+    db.query(models.Settings).filter(models.Settings.conference_id == conference.id, models.Settings.owner_id == owner_id).delete()
+    agenda_crud.delete_agenda_by_conference_id(db, conference_id=conference.id)
+    db.delete(conference)
     db.commit()
     return True
